// frontend/src/modules/Profile/interface.ts
export interface Profile {
  id: number
  avatar_url: string
  bio: string
  name: string
  login: string
  followers: number
  following: number
  location: string
  email: string
  blog: string
  twitter_username: string
}

export interface Repository {
  id: number
  name: string
  fork: boolean
  description: string
  topics: string[]
  language: Language
  homepage: string
  stargazers_count: number
  pushed_at: string
}

export type Language =
  | 'javascript'
  | 'typescript'
  | 'css'
  | 'html'
  | 'python'
  | 'vue'
  | 'ruby'